added overwrite option for Copy/Move files

+++ file.c
@@ -482,7 +482,11 @@ int copyPath(const char *src_path, const char *dst_path, FileProcessParam *param
         if (SCE_S_ISDIR(dir.d_stat.st_mode)) {
           ret = copyPath(new_src_path, new_dst_path, param);
         } else {
-          ret = copyFile(new_src_path, new_dst_path, param);
+			if (!vitashell_config.overwrite_files && checkFileExist(new_dst_path)) {
+				ret = copyFile(new_src_path, getFileNameUpdated(new_dst_path), param);
+			} else {
+				ret = copyFile(new_src_path, new_dst_path, param);
+			}
         }
 
         free(new_dst_path);
@@ -497,7 +501,11 @@ int copyPath(const char *src_path, const char *dst_path, FileProcessParam *param
 
     sceIoDclose(dfd);
   } else {
-    return copyFile(src_path, dst_path, param);
+	if (!vitashell_config.overwrite_files && checkFileExist(dst_path)) {
+		return copyFile(src_path, getFileNameUpdated(dst_path), param);
+	} else {
+		return copyFile(src_path, dst_path, param);
+	}
   }
 
   return 1;
@@ -542,12 +550,19 @@ int movePath(const char *src_path, const char *dst_path, int flags, FileProcessP
 
     // Replace file
     if (!src_is_dir && !dst_is_dir && flags & MOVE_REPLACE) {
-      sceIoRemove(dst_path);
-
-      res = sceIoRename(src_path, dst_path);
-      if (res < 0)
-        return res;
-
+      
+	  if (vitashell_config.overwrite_files)
+		  sceIoRemove(dst_path);
+	  
+	  if (!vitashell_config.overwrite_files && checkFileExist(dst_path)) {
+		  res = sceIoRename(src_path, getFileNameUpdated(dst_path));
+	  } else {
+		  res = sceIoRename(src_path, dst_path);
+	  }
+	  
+	  if (res < 0)
+		  return res;
+		 
       return 1;
     }
 
@@ -572,8 +587,14 @@ int movePath(const char *src_path, const char *dst_path, int flags, FileProcessP
           snprintf(new_dst_path, MAX_PATH_LENGTH, "%s%s%s", dst_path, hasEndSlash(dst_path) ? "" : "/", dir.d_name);
 
           // Recursive move
-          int ret = movePath(new_src_path, new_dst_path, flags, param);
-
+          //int ret = movePath(new_src_path, new_dst_path, flags, param);
+			int ret;
+			if (!vitashell_config.overwrite_files && checkFileExist(new_dst_path)) {
+				ret = movePath(new_src_path, getFileNameUpdated(new_dst_path), flags, param);
+			} else {
+				ret = movePath(new_src_path, new_dst_path, flags, param);
+			}
+			
           free(new_dst_path);
           free(new_src_path);
